+++ saleor/plugins/webhook/tasks.py
@@ -465,7 +465,7 @@ def send_webhook_using_scheme_method(
     retry_backoff=10,
     retry_kwargs={"max_retries": 5},
 )
-def send_webhook_request_async(self, event_delivery_id):
+def send_webhook_request_async(self, event_delivery_id, clear=True):
     try:
         delivery = EventDelivery.objects.select_related("payload", "webhook__app").get(
             id=event_delivery_id
@@ -499,6 +499,7 @@ def send_webhook_request_async(self, event_delivery_id):
                 delivery.event_type,
                 data,
             )
+
         attempt_update(attempt, response)
         if response.status == EventDeliveryStatus.FAILED:
             task_logger.info(
@@ -540,7 +541,8 @@ def send_webhook_request_async(self, event_delivery_id):
         attempt_update(attempt, response)
         delivery_update(delivery=delivery, status=EventDeliveryStatus.FAILED)
     observability.report_event_delivery_attempt(attempt)
-    clear_successful_delivery(delivery)
+    if clear:
+        clear_successful_delivery(delivery)
 
 
 def send_webhook_request_sync(