+++ CHANGELOG.md
@@ -35,6 +35,8 @@ All notable, unreleased changes to this project will be documented in this file.
   - `Order`
   - `User`
   - `Warehouse`
+- Add `WebhookDryRun` mutation - #11548 by @zedzior
+- Add `WebhookTrigger` mutation - #11687 by @zedzior
 
 ### Other changes